src/main/java/de/innovationhub/prox/modules/user/application/lecturer/usecase/queries/FindAllLecturersHandler.java
package de.innovationhub.prox.modules.user.application.lecturer.usecase.queries;

import de.innovationhub.prox.modules.commons.application.ApplicationComponent;
import de.innovationhub.prox.modules.user.domain.lecturer.LecturerProfile;
import de.innovationhub.prox.modules.user.domain.lecturer.LecturerProfileRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;

@ApplicationComponent
@RequiredArgsConstructor
public class FindAllLecturersHandler {

  private final LecturerProfileRepository lecturerRepository;

  public Page<LecturerProfile> handle(Pageable pageable) {
    return lecturerRepository.findAll(pageable);
  }
}
